requirements.txt
Django>=4.1
djangorestframework>=3.14
drf-yasg==1.21.8
gunicorn==20.1.0
psycopg2-binary==2.9.10
django-split-settings==1.2.0
django-cors-headers==4.5.0